push up management of activeUid state and pass isActive prop down
