+++ editor/modes/visual-editor/block.js
@@ -3,7 +3,7 @@
  */
 import { connect } from 'react-redux';
 
-function VisualEditorBlock( { uid, block, onChange } ) {
+function VisualEditorBlock( { uid, block, isActive, onChange, onFocus } ) {
 	const settings = wp.blocks.getBlockSettings( block.blockType );
 
 	let BlockEdit;
@@ -25,24 +25,35 @@ function VisualEditorBlock( { uid, block, onChange } ) {
 	}
 
 	return (
-		<BlockEdit
-			uid={uid}
-			attributes={ block.attributes }
-			onChange={ onAttributesChange } />
+		<div role="none" onFocus={ onFocus } >
+			<BlockEdit
+				uid={ uid }
+				isActive={ isActive }
+				attributes={ block.attributes }
+				onChange={ onAttributesChange } />
+		</div>
 	);
 }
 
-export default connect(
-	( state, ownProps ) => ( {
-		block: state.blocks.byUid[ ownProps.uid ]
-	} ),
-	( dispatch, ownProps ) => ( {
-		onChange( updates ) {
-			dispatch( {
-				type: 'UPDATE_BLOCK',
-				uid: ownProps.uid,
-				updates
-			} );
-		}
-	} )
-)( VisualEditorBlock );
+const mapStateToProps = ( state, ownProps ) => ( {
+	block: state.blocks.byUid[ ownProps.uid ],
+	isActive: ownProps.uid === state.blocks.activeUid
+} );
+
+const mapDispatchToProps = ( dispatch, ownProps ) => ( {
+	onChange( updates ) {
+		dispatch( {
+			type: 'UPDATE_BLOCK',
+			uid: ownProps.uid,
+			updates
+		} );
+	},
+	onFocus( ) {
+		dispatch( {
+			type: 'ACTIVE_BLOCK',
+			uid: ownProps.uid
+		} );
+	}
+} );
+
+export default connect( mapStateToProps, mapDispatchToProps )( VisualEditorBlock );
